Improve cmake configs
Daniel Pfeifer advises against intrusive valgrind and instrumentation
testing: https://youtu.be/bsXLMQ6WgIk?t=3487
He says that such testing must be external towards the project itself.
This means that we have to add coverage compile flags directory-wise,
without distinguishing between first- and third- party targets.
Of course, we should not add such flags to third-party

In the same time, Mateusz Pusz says, that it is not recommended to add
third-party with add_subdirectory, but rather use package manager:
https://youtu.be/S4QSKLXdTtA?t=221
Doing so, it is possible to safely add such flags directory-wise.
That way, code coverage will be enabled unintrusively, except for the
couple of lines in the root CMakeLists.txt