Revert "Remove intrusive target_enable_instrumentation function"
This reverts commit 235e75c7092822d404616152877175c3a7e4eeda.

Continuing reasoning about Daniel Pfeifer's “Effective CMake" and
Mateusz Pusz' "Git, CMake, Conan", making code_coverage completely
non-intrusive have several disadvantages.

There are two ways to implement it:
a. With add_compiler_options
b. With listing all targets and trying to guess which are non-imported
  non-interface and assigning them compiler options on per-target basis

Disadvantages are:
1.a. Approach (a) just breaks modern cmake idea of "targets and properties"
1.b. Approach (b) does the same but in more ugly and even more
  unstable way
1. Consequently, both do not allow to control code_coverage and other
   instrumentation on per-target basis.
2. These approaches allow to make code_coverage non-intrusive ONLY for
   exported targets, but not tests. This is because project-agnostic
   system does not really know how to properly run lit tests, or even
   how to use llvm code_coverage.

Daniel Pfeifer says that it is not desired to have copy of same tests
with code_coverage or valgrind.
Question: is it OK to have only one copy of such tests and supporting
infrastructure, which allow to optionally run them with code_coverage or
valgrind? If yes, then probably I have one.